+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 1.3.0
 
+#### New features and improvements
+
+- DatePicker:
+  - Add new method `setDateFormat(String format)` to format date (issues #5)
+
 #### Fixed bugs
 
 - PanelSlider `flush` image after end of animation
@@ -11,7 +16,7 @@
 - DatePicker:
   - Add new method `selectCurrentMonth()` (PR #2)
 - TimePicker:
-  - Add new method `setColor(Color color)` (issue #3)
+  - Add new method `setColor(Color color)` (issues #3)
 - Popup menu will show inside the windows frame
 - Add new method `setEditorIcon(Icon icon)`
 
@@ -22,7 +27,7 @@
   - Fixed `between date` resend value when editor value changed
 - TimePicker: 
   - Fixed event time changed not work while `am` or `pm` changed in editor
-  - Fixed time format by use `Locale.ENGLISH` (issue #4)
+  - Fixed time format by use `Locale.ENGLISH` (issues #4)
 - Fixed reset old editor to default after changed new editor
 - Fixed event `SelectionListener` invoke only value changed